Actualizacion empleado Service

+++ src/main/java/com/example/horarios/Repositorio/EmpleadoRepositorio.java
@@ -5,6 +5,6 @@ import com.example.horarios.Entidades.Empleado;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface EmpleadoRepositorio extends JpaRepository<Empleado, Integer> {
+public interface EmpleadoRepositorio extends JpaRepository<Empleado, Long> {
     
 }
